Converting an Object of class amelia

This function converts a list of multiply imputed data sets to an object of class amelia.

datlist2Amelia(datlist)

Arguments

datlist
List of multiply imputed data sets or an object of class mids

Value

An object of class amelia

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: Imputation of NHANES data using mice package
#############################################################################

library(mice)
library(Amelia)

data(nhanes,package="mice")
set.seed(566)  # fix random seed

# imputation with mice
imp <- mice::mice(nhanes, m=7)

# conversion to amelia object
amp <- miceadds::datlist2Amelia(datlist=imp)
str(amp)
plot(amp)
print(amp)
summary(amp)
}