Some Additional Multiple Imputation Functions, Especially for 'mice'

Contains functions for multiple imputation which complements existing functionality in R. In particular, several imputation methods for the mice package (van Buuren & Groothuis-Oudshoorn, 2011, <doi:10.18637/jss.v045.i03>) are included. Main features of the miceadds package include plausible value imputation (Mislevy, 1991, <doi:10.1007/BF02294457>), multilevel imputation for variables at any level or with any number of hierarchical and non-hierarchical levels (Grund, Luedtke & Robitzsch, 2018, <doi:10.1177/1094428117703686>; van Buuren, 2018, Ch.7, <doi:10.1201/9780429492259>), imputation using partial least squares (PLS) for high dimensional predictors (Robitzsch, Pham & Yanagida, 2016), nested multiple imputation (Rubin, 2003, <doi:10.1111/1467-9574.00217>) and substantive model compatible imputation (Bartlett et al., 2015, <doi:10.1177/0962280214521348>).

Details

The miceadds package contains some functionality for imputation of multilevel data. The function mice.impute.ml.lmer is a general function for imputing multilevel data with hierarchical or cross-classified structures for variables at an arbitrary level. This imputation method uses the lme4::lmer function in the lme4 package. The imputation method mice.impute.2lonly.function conducts an imputation for a variable at a higher level for already defined imputation methods in the mice package. Two-level imputation is available in several functions in the mice package (mice::mice.impute.2l.pan, mice::mice.impute.2l.norm) as well in micemd and hmi packages. The miceadds package contains additional imputation methods for two-level datasets: mice.impute.2l.continuous for normally distributed data, mice.impute.2l.pmm for predictive mean matching in multilevel models and mice.impute.2l.binary for binary data.

In addition to the usual mice imputation function which employs parallel chains, the function mice.1chain does multiple imputation from a single chain.

Nested multiple imputation can be conducted with mice.nmi. The function NMIcombine conducts statistical inference for nested multiply imputed datasets.

Imputation based on partial least squares regression is implemented in mice.impute.pls.

Unidimensional plausible value imputation for latent variables (or variables with measurement error) in the mice sequential imputation framework can be applied by using the method mice.impute.plausible.values.

The method mice.impute.simputation is a wrapper function to imputation methods in the simputation package. The methods mice.impute.imputeR.lmFun and mice.impute.imputeR.cFun are wrapper functions to imputation methods in the imputeR package.

The miceadds package also includes some functions R utility functions (e.g. write.pspp, ma.scale2).

Imputations for questionnaire items can be accomplished by two-way imputation (tw.imputation).

References

Bartlett, J. W., Seaman, S. R., White, I. R., Carpenter, J. R., & Alzheimer's Disease Neuroimaging Initiative (2015). Multiple imputation of covariates by fully conditional specification: Accommodating the substantive model. Statistical Methods in Medical Research, 24(4), 462-487. doi: 10.1177/0962280214521348

Grund, S., Luedtke, O., & Robitzsch, A. (2018). Multiple imputation of multilevel data in organizational research. Organizational Research Methods, 21(1), 111-149. doi: 10.1177/1094428117703686

Mislevy, R. J. (1991). Randomization-based inference about latent variables from complex samples. Psychometrika, 56(2), 177-196. doi: 10.1007/BF02294457

Robitzsch, A., Pham, G., & Yanagida, T. (2016). Fehlende Daten und Plausible Values. In S. Breit & C. Schreiner (Hrsg.). Large-Scale Assessment mit R: Methodische Grundlagen der oesterreichischen Bildungsstandardueberpruefung (S. 259-293). Wien: facultas.

Rubin, D. B. (2003). Nested multiple imputation of NMES via partially incompatible MCMC. Statistica Neerlandica, 57(1), 3-18. doi: 10.1111/1467-9574.00217

van Buuren, S. (2018). Flexible imputation of missing data. Boca Raton: CRC Press. doi: 10.1201/9780429492259

van Buuren, S., & Groothuis-Oudshoorn, K. (2011). mice: Multivariate imputation by chained equations in R. Journal of Statistical Software, 45(3), 1-67. doi: 10.18637/jss.v045.i03

See also

Examples

##
##   ::'''''''''''''''''''''''''''''''''::
##   :: miceadds 0.11-69 (2013-12-01)   ::
##   ::'''''''''''''''''''''''''''''''''::
##
##  ----------------------- mice at work ---------------------------------
##
##                         (\-.
##                         / _`> .---------.
##                 _)     / _)=  |'-------'|
##                (      / _/    |O   O   o|
##                 `-.__(___)_   | o O . o |
##                               `---------'
##
##                                          oo__
##                                         <;___)------
##                                    oo__   " "
##                                   <;___)------     oo__
##                                     " "           <;___)------
##                                                     " "