Combination of Chi Square Statistics of Multiply Imputed Datasets

This function does inference for the \(\chi^2\) statistic based on multiply imputed datasets (see e.g. Enders, 2010, p. 239 ff.; Allison, 2002). This function is also denoted as the \(D_2\) statistic.

micombine.chisquare(dk, df, display=TRUE, version=1)

Arguments

dk
Vector of chi square statistics

df
Degrees of freedom of \(\chi^2\) statistic

display
An optional logical indicating whether results should be printed at the R console.

version
Integer indicating which calculation formula should be used. The default version=1 refers to the correct formula as in Enders (2010), while version=0 uses an incorrect formula as printed in Allison (2001). The incorrect calculation version=0 was included in miceadds versions smaller than version 2.0. See also http://statisticalhorizons.com/wp-content/uploads/2012/01/combchi.sas.

Value

A vector with following entries

D

Combined \(D_2\) statistic which is approximately \(F\)-distributed with (df, df2) degrees of freedom

p

The p value corresponding to D

df

Numerator degrees of freedom

df2

Denominator degrees of freedom

References

Allison, P. D. (2002). Missing data. Newbury Park, CA: Sage.

Enders, C. K. (2010). Applied missing data analysis. Guilford Press.

See also

Examples

#############################################################################
# EXAMPLE 1: Chi square values of analyses from 7 multiply imputed datasets
#############################################################################

# Vector of 7 chi square statistics
dk <- c(24.957, 18.051, 18.812, 17.362, 21.234, 18.615, 19.84)
dk.comb <- miceadds::micombine.chisquare(dk=dk, df=4 )
  ##  Combination of Chi Square Statistics for Multiply Imputed Data
  ##  Using 7 Imputed Data Sets
  ##  F(4, 482.06)=4.438     p=0.00157