Synthesizing Method for Fixed Values by Design in synthpop

Defines a synthesizing method for fixed values of a variable by design in the synthpop package.

syn.constant(y, x, xp, fixed_values, ...)

Arguments

y

Original data vector of length \(n\)

x

Matrix (\(n \times p\)) of original covariates

xp

Matrix (\(k \times p\)) of synthesised covariates

fixed_values

Vector containing fixed values

...

Further arguments to be passed

Details

When using the synthesis method "mice" in synthpop::syn, the function argument has to appear as rf.fixed_values (convention in synthpop).

Value

A vector of length k with synthetic values of y.

See also

Examples

if (FALSE) {
#############################################################################
# EXAMPLE 1: SD2011 | Fixed values for variable sex
#############################################################################

library(synthpop)

#** selection of dataset
data(SD2011, package="synthpop")
vars <- c("sex","age","ls","smoke")
dat  <- SD2011[1:1000, vars]
dat$ls <- as.numeric(dat$ls)

#** default synthesis
imp0 <- synthpop::syn(dat)
pred <- imp0$predictor.matrix
method <- imp0$method

#** constant vector
method["sex"] <- "constant"
fixed_values <- data.frame( sex=rep(dat$sex[c(1,2)], each=1000) )
imp <- synthpop::syn( dat, method=method, k=2000, m=1,
                rf.fixed_values=fixed_values)
table(imp$syn$sex)
}