Example Datasets for miceadds Package

Example datasets for miceadds package.

data(data.ma01)
data(data.ma02)
data(data.ma03)
data(data.ma04)
data(data.ma05)
data(data.ma06)
data(data.ma07)
data(data.ma08)

Format

Dataset data.ma01:

Dataset with students nested within school and student weights (studwgt). The format is

'data.frame': 4073 obs. of 11 variables:
$ idstud : num 1e+07 1e+07 1e+07 1e+07 1e+07 ...
$ idschool: num 1001 1001 1001 1001 1001 ...
$ studwgt : num 6.05 6.05 5.27 5.27 6.05 ...
$ math : int 594 605 616 524 685 387 536 594 387 562 ...
$ read : int 647 651 539 551 689 502 503 597 580 576 ...
$ migrant : int 0 0 0 1 0 0 1 0 0 0 ...
$ books : int 6 6 5 2 6 3 4 6 6 5 ...
$ hisei : int NA 77 69 45 66 53 43 NA 64 50 ...
$ paredu : int 3 7 7 2 7 3 4 NA 7 3 ...
$ female : int 1 1 0 0 1 1 0 0 1 1 ...
$ urban : num 1 1 1 1 1 1 1 1 1 1 ...

Dataset data.ma02:

10 multiply imputed datasets of incomplete data data.ma01. The format is

List of 10
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:
$ :'data.frame': 4073 obs. of 11 variables:

Dataset data.ma03:

This dataset contains one variable math_EAP for which a conditional posterior distribution with EAP and its associated standard deviation is available.

'data.frame': 120 obs. of 8 variables:
$ idstud : int 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 ...
$ female : int 0 1 1 1 1 0 1 1 1 1 ...
$ migrant : int 1 1 0 1 1 0 0 0 1 0 ...
$ hisei : int 44 NA 26 NA 32 60 31 NA 34 26 ...
$ educ : int NA 2 NA 1 4 NA 2 NA 2 NA ...
$ read_wle : num 74.8 78.1 103.2 81.2 119.2 ...
$ math_EAP : num 337 342 264 285 420 ...
$ math_SEEAP: num 28 29.5 28.6 28.5 27.5 ...

Dataset data.ma04:

This dataset contains two hypothetical scales A and B and single variables V5, V6 and V7.

'data.frame': 281 obs. of 13 variables:
$ group: int 1 1 1 1 1 1 1 1 1 1 ...
$ A1 : int 2 2 2 1 1 3 3 NA 2 1 ...
$ A2 : int 2 2 2 3 1 2 4 4 4 4 ...
$ A3 : int 2 3 3 4 1 3 2 2 2 4 ...
$ A4 : int 3 4 6 4 7 5 3 5 5 1 ...
$ V5 : int 2 2 5 5 4 3 4 1 3 4 ...
$ V6 : int 2 5 5 1 1 3 2 2 2 4 ...
$ V7 : int 6 NA 4 5 6 2 5 5 6 7 ...
$ B1 : int 7 NA 6 4 5 2 5 7 3 7 ...
$ B2 : int 6 NA NA 6 3 3 4 6 6 7 ...
$ B3 : int 7 NA 7 4 3 4 3 7 5 NA ...
$ B4 : int 4 5 6 5 4 3 4 5 2 1 ...
$ B5 : int 7 NA 7 4 4 3 5 7 5 4 ...

Dataset data.ma05:

This is a two-level dataset with students nested within classes. Variables at the student level are Dscore, Mscore, denote, manote, misei and migrant. Variables at the class level are sprengel and groesse.

'data.frame': 1673 obs. of 10 variables:
$ idstud : int 100110001 100110002 100110003 100110004 100110005 ...
$ idclass : int 1001 1001 1001 1001 1001 1001 1001 1001 1001 1001 ...
$ Dscore : int NA 558 643 611 518 552 NA 534 409 543 ...
$ Mscore : int 404 563 569 621 653 651 510 NA 517 566 ...
$ denote : int NA 1 1 1 3 2 3 2 3 2 ...
$ manote : int NA 1 1 1 1 1 2 2 2 1 ...
$ misei : int NA 51 NA 38 NA 50 53 53 38 NA ...
$ migrant : int NA 0 0 NA 0 0 0 0 0 NA ...
$ sprengel: int 0 0 0 0 0 0 0 0 0 0 ...
$ groesse : int 25 25 25 25 25 25 25 25 25 25 ...

Dataset data.ma06:

This is a dataset in which the variable FC is only available with grouped values (coarse data or interval data).

'data.frame': 198 obs. of 7 variables:
$ id : num 1001 1002 1003 1004 1005 ...
$ A1 : int 14 7 10 15 0 5 9 6 8 0 ...
$ A2 : int 5 6 4 8 2 5 4 0 7 0 ...
$ Edu : int 4 3 1 5 5 1 NA 1 5 3 ...
$ FC : int 3 2 2 2 2 NA NA 2 2 NA ...
$ FC_low: num 10 5 5 5 5 0 0 5 5 0 ...
$ FC_upp: num 15 10 10 10 10 100 100 10 10 100 ...

Dataset data.ma07:

This is a three-level dataset in which the variable FC is only available with grouped values (coarse data or interval data).

'data.frame': 1600 obs. of 9 variables:
$ id3: num 1001 1001 1001 1001 1001 ...
$ id2: num 101 101 101 101 101 101 101 101 101 101 ...
$ id1: int 1 2 3 4 5 6 7 8 9 10 ...
$ x1 : num 0.91 1.88 NA 1.52 0.93 0.51 2.11 0.99 2.42 NA ...
$ x2 : num -0.58 1.12 0.87 -0.01 -0.14 0.48 1.85 -0.9 0.93 0.63 ...
$ y1 : num 1.66 1.66 1.66 1.66 1.66 1.66 1.66 1.66 1.66 1.66 ...
$ y2 : num 0.96 0.96 0.96 0.96 0.96 0.96 0.96 0.96 0.96 0.96 ...
$ z1 : num -0.53 -0.53 -0.53 -0.53 -0.53 -0.53 -0.53 -0.53 -0.53 -0.53 ...
$ z2 : num 0.42 0.42 0.42 0.42 0.42 0.42 0.42 0.42 0.42 0.42 ...

Dataset data.ma08:

List with several vector of strings containing descriptive data from published articles. See string_to_matrix for converting these strings into matrices.

List of 4
$ mat1: chr [1:6] "1. T1_mental_health" ...
$ mat2: chr [1:16] "1. Exp voc-T1 -" ...
$ mat3: chr [1:12] "1. TOWRE age 7\t-\t\t\t\t\t\t" ...
$ mat4: chr [1:18] "1. Vocab. age 7\t-\t\t\t\t\t" ...

Dataset data.ma09:

This is a subset of a PISA dataset that is used for generating synthetic data.

'data.frame': 342 obs. of 41 variables:
$ SEX : int 1 2 1 2 1 2 2 2 2 1 ...
$ AGE : num 16 15.9 16.3 15.5 15.9 ...
$ HISEI : int 37 46 66 51 25 NA 54 52 51 69 ...
$ FISCED : int 3 3 6 3 3 NA 3 3 2 2 ...
$ MISCED : int 3 4 4 4 3 NA 4 3 4 4 ...
$ PV1MATH: num 643 556 510 604 462 ...
$ M474Q01: int 1 1 1 1 0 1 1 1 1 0 ...
$ M155Q02: int 2 2 2 2 2 0 0 2 2 2 ...
$ M155Q01: int 1 1 0 1 1 1 1 1 1 1 ...
[...]