Combination of F Statistics for Multiply Imputed Datasets Using a Chi Square Approximation

Several \(F\) statistics from multiply imputed datasets are combined using an approximation based on \(\chi^2\) statistics (see micombine.chisquare).

micombine.F(Fvalues, df1, display=TRUE, version=1)

Arguments

Fvalues

Vector containing \(F\) values.

df1

Degrees of freedom of the numerator. Degrees of freedom of the numerator are approximated by \(\infty\) (large number of degrees of freedom).

display

A logical indicating whether results should be displayed at the console

version

Integer indicating which calculation formula should be used. The default version=1 refers to the correct formula as in Enders (2010), while version=0 uses an incorrect formula as printed in Allison (2001). The incorrect calculation version=0 was included in miceadds versions smaller than version 2.0. See also http://statisticalhorizons.com/wp-content/uploads/2012/01/combchi.sas.

Value

The same output as in micombine.chisquare

References

Allison, P. D. (2002). Missing data. Newbury Park, CA: Sage.

Enders, C. K. (2010). Applied missing data analysis. Guilford Press.

Grund, S., Luedtke, O., & Robitzsch, A. (2016). Pooling ANOVA results from multiply imputed datasets: A simulation study. Methodology, 12(3), 75-88. doi:10.1027/1614-2241/a000111

See also

Examples

#############################################################################
# EXAMPLE 1: F statistics for 5 imputed datasets
#############################################################################

Fvalues <- c( 6.76, 4.54, 4.23, 5.45, 4.78 )
micombine.F(Fvalues, df1=4 )
  ##  Combination of Chi Square Statistics for Multiply Imputed Data
  ##  Using 5 Imputed Data Sets
  ##  F(4, 52.94)=3.946     p=0.00709