Export mids object to MLwiN

Converts a mids object into a format recognized by the multilevel software MLwiN.

mids2mlwin(imp, file.prefix, path=getwd(), sep=" ", dec=".", silent=FALSE,
   X=NULL)

Arguments

imp
The imp argument is an object of class mids, typically produced by the mice() function.

file.prefix
A character string describing the prefix of the output data files.

path
A character string containing the path of the output file. By default, files are written to the current R working directory.

sep
The separator between the data fields.

dec
The decimal separator for numerical data.

silent
A logical flag stating whether the names of the files should be printed.

X
Optional data frame of variables to be included in imputed datasets.

Value

The return value is NULL.

Examples

if (FALSE) {
# imputation nhanes data
data(nhanes)
imp <- mice::mice(nhanes)
# write files to MLwiN
mids2mlwin(imp, file.prefix="nhanes" )
}